requirements.txt
appdirs==1.4.4
asgiref==3.4.1
blinker==1.4
Brlapi==0.8.2
cachetools==4.2.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info==1.0
Django==3.2.5
django-active-link==0.1.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
filelock==3.0.12
google-auth==1.32.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
gunicorn==20.1.0
gyp==0.1
hidpidaemon==18.4.6
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
jeepney==0.6.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.16.0
macaroonbakery==1.3.1
mock==4.0.3
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.0
pbr==5.5.1
Pillow==8.3.1
pop-transition==1.1.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-debian==0.1.39
python-xlib==0.29
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
QtPy==1.9.0
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
SecretStorage==3.3.1
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
systemd-python==234
twitter==1.19.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.26.2
virtualenv==20.4.0+ds
wadllib==1.3.5
whitenoise==5.3.0
xdg==5
xkit==0.0.0
zipp==1.0.0
